Babel: add RFC9229 (v4 via v6) support
This implements [RFC9229] an IPv4 via IPv6 extension
to the Babel routing protocol that allows annoncing routes to an IPv4
prefix with an IPv6 next-hop, which makes it possible for IPv4 traffic
to flow through interfaces that have not been assigned an IPv4 address.

The implementation is compatible with the current Babeld version (the
relevant changes can be seen in the [babeld PR]). I've verified this
with a few VMs in the following setup:

Bird <- v4 only -> Bird <- v6 only -> Babeld <- v4 only -> Babeld

Each routing daemon was running on their own VM and had L2 connectivity
to only its direct neighbors. At the nodes at the edges v4 networks have
been announced and full end-to-end communication was possible over the
mixed AF network. The v6 only link between Babel and Bird (at the
"center" of the above setup) did transport the v4 packets via the v6
link-local next hop addresses just as expected.

Thanks to Toke Høiland-Jørgensen for early review on this work.

[RFC9229]: https://www.ietf.org/rfc/rfc9229.txt
[babeld PR]: https://github.com/jech/babeld/pull/56

Acked-by: Toke Høiland-Jørgensen <toke@toke.dk>